(BSR)[BO] chore: rename form.py into forms.py to be consistent
