+++ api/src/pcapi/routes/backoffice/bookings/collective_bookings_blueprint.py
@@ -21,7 +21,7 @@ from pcapi.core.offerers import models as offerers_models
 from pcapi.core.permissions import models as perm_models
 from pcapi.routes.backoffice import autocomplete
 from pcapi.routes.backoffice import utils
-from pcapi.routes.backoffice.bookings import form as booking_forms
+from pcapi.routes.backoffice.bookings import forms as booking_forms
 from pcapi.routes.backoffice.bookings import helpers as booking_helpers
 from pcapi.routes.backoffice.forms import empty as empty_forms
