+++ api/src/pcapi/routes/backoffice/bookings/helpers.py
@@ -10,7 +10,7 @@ from pcapi.core.educational import models as educational_models
 from pcapi.core.finance import models as finance_models
 from pcapi.core.offers import models as offers_models
 from pcapi.core.users import models as users_models
-from pcapi.routes.backoffice.bookings.form import BaseBookingListForm
+from pcapi.routes.backoffice.bookings.forms import BaseBookingListForm
 from pcapi.utils import date as date_utils
 from pcapi.utils import email as email_utils
